Revise file structure
